+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Options screen with option to enable sounds
 ### Fixed
 - Fixed slow window resizing
+- Sounds playing when preset is loaded
 
 ## [1.0.0] - 2020-05-28
 ### Added